// src/app/_shared/interfaces/employee.ts
export interface Employee {
    id: string;
    name: string;
    lastName: string;
    dni: string;
    email: string;
    phone: string;
    city: string;
    technology: string;
    experience: Date;
    isRemote: boolean;
    salary: number;
    position: string;
    seniority: Date;
    ssNumber: string;
    iban: string;
    linkedIn: string;
    noticePeriod: number;
    comments: string;
}